Improve inlining of critical functions of TaskQueue (#442)
* Implement TF_FORCE_INLINE and TF_NO_INLINE

* Split TaskQueue::push() into fast and slow paths

This function is part of the hot code path and contains a large part of
code that is rarely executed yet contributes to the size of function
prologue and epilogue. It makes sense to partially inline this function.

* Force-inline TaskQueue::pop()

This function is part of the hot code path and is only called from 2
places. It makes sense to inline it.